fxcalculator/requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.13.1
numpy==1.21.4
pandas==1.3.4
six==1.16.0
sqlparse==0.4.2

